split transform metafunctions into bits
