+++ include/nmtools/meta/transform.hpp
@@ -1,10 +1,30 @@
 #ifndef NMTOOLS_META_TRANSFORM_HPP
 #define NMTOOLS_META_TRANSFORM_HPP
 
+#include "nmtools/meta/bits/transform/append_type.hpp"
+#include "nmtools/meta/bits/transform/append_value.hpp"
+#include "nmtools/meta/bits/transform/at.hpp"
+#include "nmtools/meta/bits/transform/bit_reference_to_bool.hpp"
+#include "nmtools/meta/bits/transform/concat_type.hpp"
+#include "nmtools/meta/bits/transform/get_either.hpp"
+#include "nmtools/meta/bits/transform/get_element_type.hpp"
+#include "nmtools/meta/bits/transform/get_index_type.hpp"
+#include "nmtools/meta/bits/transform/get_maybe_type.hpp"
+#include "nmtools/meta/bits/transform/len.hpp"
+#include "nmtools/meta/bits/transform/promote_index.hpp"
+#include "nmtools/meta/bits/transform/promote_types.hpp"
+#include "nmtools/meta/bits/transform/remove_cvref.hpp"
+#include "nmtools/meta/bits/transform/replace_either.hpp"
+#include "nmtools/meta/bits/transform/replace_element_type.hpp"
+#include "nmtools/meta/bits/transform/resize_fixed_ndarray.hpp"
+#include "nmtools/meta/bits/transform/to_value.hpp"
+#include "nmtools/meta/bits/transform/transform_bounded_array.hpp"
+#include "nmtools/meta/bits/transform/tuple_to_array.hpp"
+#include "nmtools/meta/bits/transform/type_list_at.hpp"
+
 #include "nmtools/meta/common.hpp"
 #include "nmtools/meta/traits.hpp"
 #include "nmtools/meta/array.hpp"
-#include "nmtools/meta/arithmetic.hpp"
 
 #include <type_traits>
 #include <cassert>
@@ -15,177 +35,6 @@
 
 namespace nmtools::meta
 {
-    /**
-     * @brief transform (bounded) raw array to std::array
-     * should have member type `type` with type of std::array
-     * if given T is bounded (1D or 2D) raw array, 
-     * `type` has type of T otherwise
-     * 
-     * @tparam T 
-     */
-    template <typename T, typename=void>
-    struct transform_bounded_array 
-    {
-        using type = T;
-    };
-
-    /**
-     * @brief helper alias template to transform (bounded) raw array to std::array
-     * 
-     * @tparam T (bounded) array
-     */
-    template <typename T>
-    using transform_bounded_array_t = typename transform_bounded_array<T>::type;
-
-    /**
-     * @brief given T<Origin..>, transform to T<Substitution...>, if failed type is void
-     * variadic template parameter Origin... & Substitution... should be type parameter. 
-     * For non-type template parameter, only std::array is supported for now
-     * 
-     * @tparam T class which its template parameter to be replaced
-     * @tparam  
-     */
-    template <typename T, typename...>
-    struct replace_template_parameter 
-    {
-        // TODO: remove metafunctions
-        using origin_tparams = void;
-        /**
-         * @brief define resulting type as void, so that
-         * replace_template_parameter_t still well-formed
-         * while error handling can be deferred to the caller
-         * hence gives much more context
-         */
-        using type = void;
-    };
-
-    // TODO: remove metafunctions
-    /**
-     * @brief specialization for replace_template_parameter for success case
-     * 
-     * @tparam T 
-     * @tparam Subs 
-     * @tparam Origin 
-     */
-    template <template<typename...> typename T, typename ...Subs, typename ...Origin>
-    struct replace_template_parameter<T<Origin...>,Subs...>
-    {
-        // use lighter type for type-list
-        using origin_tparams = std::tuple<Origin...>;
-        /**
-         * @brief this assumes that the number of given substitute type, sizeof...(Subs),
-         * is sufficient to instantiate T<Subs...>. Consider a case where this requirement
-         * may not be met, we may need to substitute the missing template parameter from Subs
-         * using appropriate type from Origin....
-         */
-        using type = T<Subs...>;
-    };
-
-    // TODO: remove metafunctions
-    /**
-     * @brief specialization for replace_template_parameter for std::array, 
-     * since we can't mix non-type & type as variadic template parameter.
-     * The new size of new array should be supplied via std::integral_constant,
-     * since we expects type here.
-     * 
-     * @tparam value_t origin value_type
-     * @tparam N origin size
-     * @tparam subs_value_t value_type for substitution
-     * @tparam subs_N substitution size
-     */
-    template <typename value_t, auto N, typename subs_value_t, auto subs_N, typename size_type>
-    struct replace_template_parameter<std::array<value_t,N>,subs_value_t,std::integral_constant<size_type,subs_N>>
-    {
-        using origin_tparams = std::tuple<value_t>;
-        using type = std::array<subs_value_t,subs_N>;
-    };
-
-    /**
-     * @brief helper alias template to replace tparam
-     * 
-     * @tparam T original type
-     * @tparam Args template parameter for substitution
-     */
-    template <typename T, typename...Args>
-    using replace_template_parameter_t = typename replace_template_parameter<T,Args...>::type;
-
-    /**
-     * @brief Type as value
-     * 
-     * @tparam T type to wrap
-     */
-    template <typename T>
-    struct as_value
-    {
-        using type = T;
-    };
-
-    /**
-     * @brief helper inline variable for as_value
-     * 
-     * @tparam T type to wrap
-     */
-    template <typename T>
-    constexpr inline auto as_value_v = as_value<T>{};
-
-    /**
-     * @brief similar to replace_template_parameter but assuming
-     * new parameter are packed in typelist (tuple). default impl
-     * has void as resulting type.
-     * 
-     * @tparam T 
-     * @tparam typename=void 
-     */
-    template <typename T, typename=void>
-    struct replace_template_parameter_from_typelist
-    {
-        /**
-         * @brief define resulting type as void, so that
-         * replace_template_parameter_t still well-formed
-         * while error handling can be deferred to the caller
-         * hence gives much more context
-         */
-        using type = void;
-    };
-
-    /**
-     * @brief replace_template_parameter of class T with parameter(s) packed as tuple.
-     *
-     * @example 
-     * @tparam T template template parameter, deduced automatically
-     * @tparam Origin template parameter(s) of T, deduced automatically
-     * @tparam Subs substitute for Origin, deduced automatically
-     */
-    template <template <typename...> typename T, typename...Origin, typename...Subs>
-    struct replace_template_parameter_from_typelist<T<Origin...>,std::tuple<Subs...>>
-    {
-        /**
-         * @brief call replace_template_parameter here so that any existing specialization
-         * can be used, e.g. specialization of std::array which unpacks integral_constant<...>
-         * to size_t as template parameter.
-         * 
-         */
-        using type = replace_template_parameter_t<T<Origin...>,Subs...>;
-    };
-
-    /**
-     * @brief helper alias template for replace_template_parameter_from_typelist.
-     * 
-     * @tparam T class which its template parameter(s) are to be replaced.
-     * @tparam TypeList type list, e.g. std::tuple<...>, which holds types for substitution.
-     */
-    template <typename T, typename TypeList>
-    using replace_template_parameter_from_typelist_t = typename replace_template_parameter_from_typelist<T,TypeList>::type;
-
-    /**
-     * @brief helper alias template to replace tparam
-     * 
-     * @tparam T original type
-     * @tparam Args template parameter for substitution
-     */
-    /* TODO: remove */
-    template <typename T, typename...Args>
-    using replace_tparam_t = typename replace_template_parameter<T,Args...>::type;
 
     namespace detail
     {
@@ -323,33 +172,6 @@ namespace nmtools::meta
     template <typename T, auto I>
     using split_at_t = type_t<split_at<T,I>>;
 
-    /*
-    TODO: consider to provide specialization for replace_template_parameter, that accepts
-        - T<typename,auto..>, 
-        - T<auto,typename...>, etc.
-    */
-
-    /**
-     * @brief resize fixed array
-     * 
-     * @tparam T original array
-     * @tparam U fixed ndarray type containing new shape
-     * @tparam typename 
-     */
-    template <typename T, typename U, typename=void>
-    struct resize_fixed_ndarray
-    {
-        using type = T;
-    }; // resize_fixed_ndarray
-
-    /**
-     * @brief helper alias template for resize fixed array
-     * 
-     * @tparam T original array
-     * @tparam U type containing new shape
-     */
-    template <typename T, typename U>
-    using resize_fixed_ndarray_t = type_t<resize_fixed_ndarray<T,U>>;
 
     template <typename T>
     struct pop_first
@@ -542,180 +364,6 @@ namespace nmtools::meta
     template <typename op_t, typename...tparams>
     using resolve_optype_t = typename resolve_optype<void,op_t,tparams...>::type;
 
-    /**
-     * @brief metafunction to get the value/element type of an ndarray
-     * 
-     * @tparam T type to test
-     * @todo consider use expr to deduce type (e.g. expr::atnd)
-     */
-    template <typename T>
-    struct get_ndarray_value_type
-    {
-        using type = void;
-    };
-
-    /**
-     * @brief helper alias template for get_ndarray_value_type
-     * 
-     * @tparam T type to test
-     */
-    template <typename T>
-    using get_ndarray_value_type_t = typename get_ndarray_value_type<T>::type;
-
-    template <typename T, typename=void>
-    struct bit_reference_to_bool
-    {
-        static constexpr auto vtype = [](){
-            if constexpr (is_bit_reference_v<T>) {
-                return as_value_v<bool>;
-            } else {
-                return as_value_v<T>;
-            }
-        }();
-        using type = type_t<decltype(vtype)>;
-    }; // bit_reference_to_bool
-
-    template <typename T>
-    using bit_reference_to_bool_t = type_t<bit_reference_to_bool<T>>;
-
-    namespace error
-    {
-        struct GET_ELEMENT_TYPE_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-
-    /**
-     * @brief 
-     * 
-     * @tparam T 
-     * @tparam typename=void 
-     */
-    template <typename T, typename=void>
-    struct get_element_type
-    {
-        /**
-         * @brief helper function to deduce resulting type
-         * 
-         * @return constexpr auto 
-         * @todo consider use expr to deduce type (e.g. expr::atnd)
-         */
-        static inline constexpr auto vtype = []()
-        {
-            /**
-             * @note since is_array1d_v and is_array2d_v may not be mutually exclusive,
-             * it is hard to use std::enable_if / std::void_t to specialize,
-             * hence using helper function instead
-             * to deduce the resulting element type.
-             * 
-             * @note using detail::void_to_fail_t since void can't be instantiated, reversed back to void at the caller site.
-             */
-            if constexpr (std::is_array_v<T>) {
-                using type = std::remove_all_extents_t<T>;
-                return as_value_v<type>;
-            }
-            else if constexpr (meta::nested_array_dim_v<T> > 0) {
-                using element_t = remove_all_nested_array_dim_t<T>;
-                // note that remove_all_nested_array_dim_t using expression to deduce the type
-                // causing vector<bool> deduced to std::_Bit_reference instead of bool
-                using type = bit_reference_to_bool_t<element_t>;
-                return as_value_v<type>;
-            }
-            // ndarray is more generic
-            else if constexpr (meta::is_ndarray_v<T>) {
-                using type = get_ndarray_value_type_t<T>;
-                return as_value_v<type>;
-            }
-            else if constexpr (meta::is_integral_constant_v<T>) {
-                return as_value_v<typename T::value_type>;
-            }
-            else if constexpr (meta::has_value_type_v<T>) {
-                using type = typename T::value_type;
-                if constexpr (std::is_arithmetic_v<type>)
-                    return as_value_v<type>;
-                else
-                    return as_value_v<typename get_element_type<type>::type>;
-            }
-            else if constexpr (std::is_arithmetic_v<T>)
-                return as_value_v<T>;
-            else return as_value_v<error::GET_ELEMENT_TYPE_UNSUPPORTED>;
-        }();
-        using type = type_t<decltype(vtype)>;
-    }; // get_element_type
-
-    /**
-     * @brief helper alias template for get_element_type
-     * 
-     * @tparam T type to transform
-     */
-    template <typename T>
-    using get_element_type_t = typename get_element_type<T>::type;
-
-    namespace detail
-    {
-        /**
-         * @brief helper metafunction for replace_element_type
-         *
-         * return void type if no specialization exists
-         * 
-         * @tparam T type in which its element type to be replaced
-         * @tparam U substitute to element type  of T
-         * @tparam typename 
-         */
-        template <typename T, typename U, typename=void>
-        struct replace_element_type_helper
-        {
-            using type = void;
-        };
-
-        // TODO: cleanup metafunctions
-        // NOTE: make it easier to recurse
-        template <typename T, typename U>
-        struct replace_element_type_helper<T,U,std::enable_if_t<std::is_arithmetic_v<U>>>
-        {
-            static constexpr auto vtype = [](){
-                if constexpr (has_value_type_v<T>) {
-                    using value_type = typename T::value_type;
-                    // TODO: wrap std::is_arithmetic using other metafunctions
-                    if constexpr (std::is_arithmetic_v<value_type>)
-                        return as_value_v<replace_template_parameter_t<T,U>>;
-                    else {
-                        using type = type_t<replace_element_type_helper<value_type,U>>;
-                        return as_value_v<replace_template_parameter_t<T,type>>;
-                    }
-                }
-                else return as_value_v<void>;
-            }();
-            using type = type_t<decltype(vtype)>;
-        };
-    } // namespace detail
-
-    /**
-     * @brief given type T, try to replace the element type of T with U.
-     *
-     * The default implementation if no specialization of type T exists is defined
-     * by helper metafunction detail::relace_element_type_helper. This helper can handles
-     * nested std::vector and std::array. By deferring implementation to helper class,
-     * this metafunction doesn't need to provide specialization hence it will be easier
-     * to specialize on custom types.
-     * 
-     * @tparam T type in which its element type to be replaced
-     * @tparam U substitute to element type  of T
-     * @tparam typename sfinae point
-     * @see detail::replace_element_type_helper
-     */
-    template <typename T, typename U, typename=void>
-    struct replace_element_type
-    {
-        using type = type_t<detail::replace_element_type_helper<T,U>>;
-    }; // replace_element_type
-
-    /**
-     * @brief helper alias template for replace element type
-     * 
-     * @tparam T type in which its element type to be replaced
-     * @tparam U substitute to element type  of T
-     */
-    template <typename T, typename U>
-    using replace_element_type_t = typename replace_element_type<T,U>::type;
 
     // TODO: cleanup metafunctions
     /**
@@ -869,29 +517,6 @@ namespace nmtools::meta
     template <auto I, auto...Is>
     using sequence_t = std::integer_sequence<decltype(I),I,Is...>;
 
-    namespace error
-    {
-        template <typename...>
-        struct TO_VALUE_UNSUPPORTED : detail::fail_t {};
-    }
-
-    /**
-     * @brief Convert constant index to value
-     * 
-     * If T is tuple, it is expected to transform to array
-     * for easy handling as value.
-     * 
-     * @tparam T 
-     */
-    template <typename T>
-    struct to_value
-    {
-        static inline constexpr auto value = error::TO_VALUE_UNSUPPORTED<>{};
-    }; // to_value
-
-    template <typename T>
-    constexpr inline auto to_value_v = to_value<T>::value;
-
     namespace detail
     {
         // TODO: remove metafunctions
@@ -954,15 +579,6 @@ namespace nmtools::meta
     template <typename array_t>
     using get_element_or_common_type_t = type_t<get_element_or_common_type<array_t>>;
 
-    template <typename T, typename=void>
-    struct tuple_to_array
-    {
-        using type = T;
-    }; // tuple_to_array
-
-    template <typename T>
-    using tuple_to_array_t = type_t<tuple_to_array<T>>;
-
     /**
      * @brief metafunction to resize fixed vector.
      * 
@@ -992,267 +608,6 @@ namespace nmtools::meta
 
     template <typename T, size_t N>
     using resize_fixed_index_array_t = type_t<resize_fixed_index_array<T,N>>;
-
-    // TODO(wrap std metafunctions): specialize on std::tuple_size_v instead of as default
-    // TODO(wrap stl): do something with std::get
-    /**
-     * @brief Helper metafunction for convinient tuple-size
-     * 
-     * @tparam T 
-     */
-    template <typename T>
-    struct len
-    {
-        static constexpr auto value = [](){
-            if constexpr (has_tuple_size_v<T>) {
-                return std::tuple_size_v<T>;
-            } else if constexpr (is_fixed_size_ndarray_v<T>) {
-                // similar to python, when len is used on ndarray
-                // return the size of first axis
-                constexpr auto shape = fixed_ndarray_shape_v<T>;
-                using shape_t = remove_cvref_t<decltype(shape)>;
-                if constexpr (has_square_bracket_v<shape_t,size_t>) {
-                    return shape[0];
-                } else if constexpr (has_bracket_v<shape_t,size_t>) {
-                    return shape(0);
-                } else if constexpr (has_template_get_v<shape_t>) {
-                    using std::get;
-                    return get<0>(shape);
-                }
-            }
-            else return 0;
-        }();
-    };
-
-    template <typename T>
-    constexpr inline auto len_v = len<T>::value;
-
-    /**
-     * @brief "value-less" type list.
-     * 
-     * @tparam Ts 
-     */
-    template <typename...Ts>
-    struct type_list {};
-
-    template <size_t I, typename type_list_t, typename=void>
-    struct type_list_at
-    {
-        using type = void;
-    }; // type_list_at
-
-    template <size_t I, typename T, typename...Ts>
-    struct type_list_at<I,type_list<T,Ts...>>
-    {
-        static constexpr auto vtype = [](){
-            if constexpr (I==0)
-                return as_value_v<T>;
-            else {
-                using new_type_list = type_list<Ts...>;
-                using type = type_t<type_list_at<I-1,new_type_list>>;
-                return as_value_v<type>;
-            }
-        }();
-        using type = type_t<decltype(vtype)>;
-    }; // type_list_at
-
-    template <size_t I, typename T>
-    using type_list_at_t = type_t<type_list_at<I,T>>;
-
-    /**
-     * @brief Helper metafunction to get type at index I from type list T.
-     * 
-     * @tparam T type list
-     * @tparam I index to get the type
-     */
-    template <typename T, size_t I>
-    struct type_at
-    {
-        using type = detail::fail_t;
-    }; // at
-
-    /**
-     * @brief Helper alias template to type_at.
-     * 
-     * @tparam T type list
-     * @tparam I index to get the type
-     */
-    template <typename T, size_t I>
-    using at_t = type_t<type_at<T,I>>;
-
-    /**
-     * @brief Append new value to the end of some value type.
-     * 
-     * @tparam value_type some value type to be appended
-     * @tparam new_value  desired new value
-     */
-    template <typename value_type, auto new_value>
-    struct append_value
-    {
-        using type = void;
-    }; // append_value
-
-    template <typename value_type, auto N>
-    using append_value_t = type_t<append_value<value_type,N>>;
-
-    namespace error
-    {
-        struct APPEND_TYPE_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-
-    /**
-     * @brief Append new type to the end of some type list.
-     * 
-     * @tparam T        some type list to be appended
-     * @tparam new_type desired new type
-     */
-    template <typename T, typename new_type>
-    struct append_type
-    {
-        using type = error::APPEND_TYPE_UNSUPPORTED;
-    }; // append_type
-
-
-    template <typename T, typename new_type>
-    using append_type_t = type_t<append_type<T,new_type>>;
-
-    namespace error
-    {
-        struct CONCAT_TYPE_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-
-    /**
-     * @brief Concatenate two type list to single type list.
-     * Given List<Ts...> and List<Us...> return List<Ts...,Us...>;
-     * 
-     * @tparam T type list
-     * @tparam U type list
-     */
-    template <typename T, typename U>
-    struct concat_type
-    {
-        using type = error::CONCAT_TYPE_UNSUPPORTED;
-    }; // concat_type
-
-    /**
-     * @brief Helper type alias for concat_type
-     * 
-     * @tparam T type list
-     * @tparam U type list
-     */
-    template <typename T, typename U>
-    using concat_type_t = type_t<concat_type<T,U>>;
-
-    namespace error
-    {
-        // default type (error-type) for promote_types metafunction
-        struct PROMOTE_TYPE_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-    
-    /**
-     * @brief Metafunction to promote / deduce the resulting type from two arrays.
-     * 
-     * @tparam tag_t can be specialize if desired.
-     * @tparam lhs_t 
-     * @tparam rhs_t 
-     * @tparam typename 
-     */
-    template <typename tag_t, typename lhs_t, typename rhs_t, typename=void>
-    struct promote_types
-    {
-        using type = error::PROMOTE_TYPE_UNSUPPORTED;
-    }; // promote_types
-
-    template <typename tag_t, typename lhs_t, typename rhs_t>
-    using promote_types_t = type_t<promote_types<tag_t,lhs_t,rhs_t>>;
-
-    namespace error
-    {
-        struct PROMOTE_INDEX_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-
-    /**
-     * @brief Metafunction to deduce common type for index operation
-     * 
-     * @tparam lhs_t 
-     * @tparam rhs_t 
-     */
-    template <typename first_t, typename second_t, typename...rest_t>
-    struct promote_index
-    {
-        template <typename left_t, typename right_t>
-        static constexpr auto non_constant(as_value<left_t>, as_value<right_t>) {
-            if constexpr (!is_constant_index_v<left_t>) {
-                return as_value_v<left_t>;
-            } else /* if constexpr (!is_constant_index_v<right_t>) */{
-                return as_value_v<right_t>;
-            }
-        }
-    
-        template <typename T>
-        static constexpr auto make_non_constant(as_value<T>) {
-            if constexpr (is_constant_index_v<T>) {
-                return as_value_v<typename T::value_type>;
-            } else {
-                return as_value_v<T>;
-            }
-        }
-
-        template <typename left_t, typename right_t>
-        static constexpr auto cast(as_value<left_t>, as_value<right_t>) {
-            // TODO: consider to select largest byte
-            if constexpr (is_signed_v<left_t>) {
-                return as_value_v<left_t>;
-            } else /* if (is_signed_v<right_t>) */ {
-                return as_value_v<right_t>;
-            }
-        }
-
-        template <typename lhs_t, typename rhs_t>
-        static constexpr auto binary_vtype(as_value<lhs_t>, as_value<rhs_t>) {
-            if constexpr (is_constant_index_v<lhs_t> && is_constant_index_v<rhs_t>) {
-                using left_t  = typename lhs_t::value_type;
-                using right_t = typename rhs_t::value_type;
-                return cast(as_value_v<left_t>, as_value_v<right_t>);
-            } else if constexpr (is_index_v<lhs_t> && is_index_v<rhs_t>) {
-                // simply select signed and non constant if possible
-                auto signed_   = cast(make_non_constant(as_value_v<lhs_t>), make_non_constant(as_value_v<rhs_t>));
-                auto non_const = make_non_constant(signed_);
-                return non_const;
-            } else {
-                return as_value_v<error::PROMOTE_INDEX_UNSUPPORTED>;
-            }
-        }
-
-        static constexpr auto vtype = [](){
-            using types = type_list<rest_t...>;
-            constexpr auto init_ = binary_vtype(as_value_v<remove_cvref_t<first_t>>,as_value_v<remove_cvref_t<second_t>>);
-            return template_reduce<sizeof...(rest_t)>([&](auto init, auto index){
-                constexpr auto i = decltype(index)::value;
-                using type_i = remove_cvref_t<at_t<types,i>>;
-                return binary_vtype(init,as_value_v<type_i>);
-            }, init_);
-        }();
-        using type = type_t<decltype(vtype)>;
-    }; // promote_index
-
-    template <typename lhs_t, typename rhs_t, typename...rest_t>
-    using promote_index_t = type_t<promote_index<lhs_t,rhs_t,rest_t...>>;
-
-    namespace error
-    {
-        // default type (error-type) for get_index_type
-        struct GET_INDEX_TYPE_UNSUPPORTED : detail::fail_t {};
-    } // namespace error
-
-    template <typename array_t, typename=void>
-    struct get_index_type
-    {
-        using type = error::GET_INDEX_TYPE_UNSUPPORTED;
-    }; // get_index_type
-
-    template <typename array_t>
-    using get_index_type_t = type_t<get_index_type<array_t>>;
 
     /**
      * @brief Alias to std::remove_pointer_t.
